BukkitClasses - cleanup and re-categorize
