+++ src/main/java/ch/njol/skript/classes/data/bukkit/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Represents {@link ch.njol.skript.classes.ClassInfo ClassInfos} relating to {@link org.bukkit.Bukkit}
+ */
+package ch.njol.skript.classes.data.bukkit;
